api: Move non-trivial middlewares to own file
